pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-computer-mcp"
version = "0.1.0"
description = "A visual computer-use MCP server for Windows and macOS."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp[cli]>=1.13.0",
  "pydantic>=2.8.0",
  "mss>=10.0.0",
  "Pillow>=10.4.0",
  "pynput>=1.7.7",
  "pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'",
  "pyobjc-framework-AppKit>=10.0; sys_platform == 'darwin'",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
vision-computer-mcp = "computer_use_mcp.server:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
